Updates for canyon (#99)
* Some refactoring for canyon development

* refactor(moth_ui): Node tree now takes a context rather than relying on
global context

- Removed the current context concept
- All nodes now store a reference to the context which gives them access
  to renderer etc.
- This allows multiple graphical contexts to work side by side.

* fix(backends): Fixed backends for the new moth_ui refactor

* fix(editor): Fix for editor with the new moth_ui refactor.

The moth context is stored on the editor layer now which supplies it to
other parts that need it.

* chore: Cleaning up include files. Removed example for now.

Tried to make the include files for moth_ui more consumer friendly. The
example project was also sorely out of date and so it's been removed for
now.

* chore: Updates as per review.

* chore: Some string parameters.

* chore: Fixing string parameters.

* fix: Possible fix for windows builds failing on github actions.

Build is falling over because of bad toolset versions. This might
address it but I aren't on a windows machine so have to commit to see if
it fixes it.

* chore: Small cleanups after testing windows build.