fix: Fixing canvas rendering. WIP
The rendering of the canvas is handled in ImGui which was commented out.
Reenabled the rendering of images through ImGui but it only works with
SDL for now. Vulkan silently crashes on the ImGui call somewhere in the
nvidia driver. Will need to look at how the texture is created/used.