refactor: Updating editor to use canyon as a backend. (#101)
* refactor: Updating editor to use canyon as a backend. WIP.

In a place where it builds and runs but does not render the canvas
properly. Both in vulkan and sdl.

* refactor: Layer stack now uses renderer to set logical rendering size.

This was moved from the old graphics context into the new rendering
interface.
Still does not seem to fix the canvas rendering issues.

* fix: Fixing canvas rendering. WIP

The rendering of the canvas is handled in ImGui which was commented out.
Reenabled the rendering of images through ImGui but it only works with
SDL for now. Vulkan silently crashes on the ImGui call somewhere in the
nvidia driver. Will need to look at how the texture is created/used.

* fix: Fixing canvas rendering in vulkan.

Turns out the issue was inside canyon. The sampler and view of the image
weren't properly being created. The fix has been applied in canyon.
Additionally ITargets are now no longer IImages but have a method to
return the underlying IImage.

* feat: Brought back example project.

Brought back the example project and all its assets. I'm mostly
interested in the assets since it was good for testing. The actual
example project running is not so important.

* chore: Updated path separators in example assets.

These were all using windows paths. They have now been updated to
generic separators. It should work on windows but is yet to be tested.

* chore: Removing todos and dead code.

* feat: Updated example project to use canyon backend.

Updated so the example app builds and runs. It works up until it loads
the next layout and crashes. This isn't a big deal for now since it's
only an example and I'm mostly using it for the assets.

* chore: Removing old VULKAN define.

* chore: Removed the backends library.

This is replaced by canyon now.

* chore: Removed space.

* fix: Updating linux CI build to use personal artifactory

* fix: Fixing github linux build. probably.

Added all the dependencies to the local conan repo so hopefully the
builds dont take so long.

* fix: Possible fix for windows. Using artifactory.

* fix: Windows editor build fix attempt.

* fix: Maybe fixed windows build. Lockfile issues.

* fix: Fixing windows build. Works locally so far.

* chore: Updating lockfile for linux.

* chore: Removing commented out stuff.

* chore: Unrestricting the cmake build requirement with a range.

* chore: Cleaning up building.

Canyon now supplies imgui so removing that from build files.
Added version config.

* fix: Making compile definition private.

* fix: Fixing version.txt not being exported for the build.

* fix: Fixing build issues on windows with MS define.